# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/maintenance/history": {
      "get": {
        "operationId": "MaintenanceController_getHistory",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance history returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceHistoryResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Get maintenance history visible to the acting role",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/maintenance/{resourceId}/accept": {
      "post": {
        "operationId": "MaintenanceController_accept",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "RES-ITL-001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance request accepted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Accept a maintenance request",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/maintenance/{resourceId}/repair": {
      "post": {
        "operationId": "MaintenanceController_repair",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "RES-ITL-001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource marked as repaired successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Mark a resource as repaired",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/maintenance/{resourceId}/scrap": {
      "post": {
        "operationId": "MaintenanceController_scrap",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "RES-ITL-001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource scrapped successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Scrap a resource from the maintenance flow",
        "tags": [
          "maintenance"
        ]
      }
    },
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_getAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "List notifications visible to the acting role",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/{id}": {
      "patch": {
        "operationId": "NotificationsController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "N1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Update a notification state",
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResourceX Demo Backend",
    "description": "Demo-exact NestJS backend for the ResourceX frontend. All authorization is role-based through request headers. Use x-user-role on every protected route and x-user-id when actor-specific scoping is required.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Required RBAC header. Example values: Requestor, Dept Head, Registrar, Staff, System Admin."
      },
      "user-id-header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Optional acting user id header for own-profile and department-scoped behavior. Example: U100."
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "object",
            "example": "Selected resources must match request quantity."
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "ResourceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "RES-1049"
          },
          "name": {
            "type": "string",
            "example": "Laptop - Dell XPS"
          },
          "type": {
            "type": "string",
            "example": "Laptop"
          },
          "department": {
            "type": "string",
            "example": "IT Services"
          },
          "serialNumber": {
            "type": "string",
            "example": "SN-998822"
          },
          "status": {
            "type": "string",
            "example": "Allocated"
          },
          "condition": {
            "type": "string",
            "example": "Good"
          },
          "assignedTo": {
            "type": "string",
            "example": "RAVI CHANDRA"
          },
          "assignedToId": {
            "type": "string",
            "example": "U1"
          },
          "date": {
            "type": "string",
            "example": "Jan 12, 2023"
          },
          "vendor": {
            "type": "string",
            "example": "Cisco Systems"
          },
          "invoice": {
            "type": "string",
            "example": "INV-8899"
          },
          "location": {
            "type": "string",
            "example": "IT Lab"
          },
          "returnRequestedBy": {
            "type": "string",
            "example": "RAVI CHANDRA"
          },
          "returnRequestedById": {
            "type": "string",
            "example": "U1"
          }
        },
        "required": [
          "id",
          "type",
          "department",
          "status"
        ]
      },
...
      "MaintenanceHistoryResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "RES-2233"
          },
          "type": {
            "type": "string",
            "example": "Printer Color"
          },
          "allocatedTo": {
            "type": "string",
            "example": "None"
          },
          "issue": {
            "type": "string",
            "example": "Paper Jam"
          },
          "actionDate": {
            "type": "string",
            "example": "Oct 16, 2023"
          },
          "status": {
            "type": "string",
            "example": "Repaired"
          },
          "department": {
            "type": "string",
            "example": "Administration"
          }
        },
        "required": [
          "code",
          "type",
          "allocatedTo",
          "issue",
          "actionDate",
          "status",
          "department"
        ]
      },
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "N1"
          },
          "title": {
            "type": "string",
            "example": "Welcome to ResourceX"
          },
          "description": {
            "type": "string",
            "example": "Your account has been created."
          },
          "time": {
            "type": "string",
            "example": "1 day ago"
          },
          "read": {
            "type": "boolean",
            "example": false
          },
          "recipientRole": {
            "type": "string",
            "example": "All"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "time",
          "read",
          "recipientRole"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "read": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}